#343 refactor: place-review 도메인 의존성 분리 및 PlaceEmbeddingFacade 도입
- RetrievalTextBuilder, PlaceEmbeddingBatchProcessor 파라미터를 PlaceReviewSummary 엔티티 → String reviewSummary로 변경
- PlaceSearchDocumentService에서 review 의존성 및 배치 로직 제거, DIRTY 상태 관리만 담당
- PlaceEmbeddingFacade 신규 생성: 임베딩 실행 오케스트레이션 및 review 레포지토리 의존성 집중
- AdminPlaceFacade에서 초기화 호출을 PlaceEmbeddingFacade로 위임
